Pull common code from py impls out
